Fix Kernel version regular expression in Settings
The kernel version string from /proc/version looks like the
following:

 Linux version 3.4.0-aia1+ (user@sysname) (gcc version 4.6
 20110718 (prerelease) (GCC) ) #1 SMP Fri May 25 07:26:49
 PDT 2012

The regular expression used by the Settings application to
parse this had several problems:

 () the part for recognizing the "(user@sysname)" part
    allowed itself to consume the trailing ")" as part of
    the sysname if the sysname was not fully qualified
    (i.e., containing ".").
 () the part for skipping parsing "(gcc ...)" had an
    ambiguous repetition specifier that allowed the second
    nested parentheses group to terminate the entire "(gcc
    ...)" section, leaving one "un-eaten" ")" which then
    took the place of the "group 3" part.  Symptom:  the
    "user@sysname" part was displayed as "user@sysname) )".

 () The regular expression only dropped a "PREEMPT" keyword
    after the "#1" part, instead of both "PREEMPT" and "SMP"
    keywords.  Symptom:  the "SMP" keyword was erroneously
    printed.

This change alters the sysname recognizer so that a ")"
properly terminates it.  The gcc-skipper part of the regular
expression is altered to be of the form
"(leader(stuff)trailer)" with the "(stuff)trailer" part
allowed to repeat zero or more times and suitable
termination rules for "leader", "stuff", and "trailer".  It
also adds the "SMP" keyword to the set of skipped keywords.

Issue: AIA-501
Change-Id: Ib9aa0775454de1a59af23565057bed923c170450
Signed-off-by: Paul Drews <paul.drews@intel.com>