Reflect Actual Screen Orientation in the Screen Orientation List
For devices without accelerometers, an option is provided in
Settings->Display for user to set screen orientation.

This change updates the list according to the actual screen
orientation at the time Settings app is launched. This is to
fix an issue where some apps change the screen orientation but
the list is still highlighting an old setting.

Issue: AXIA-1252
Change-Id: I419a9b8cf58baddef683882c9f159aec95d7007a
Signed-off-by: Daniel Leung <daniel.leung@intel.com>