Make automatic brightness control depend on FEATURE_SENSOR_LIGHT
The system defines several hardware/product build values in
frameworks/base/core/res/res/values/config.xml which are dynamically
checked in frameworks to provide the appropriate experience and
functionality.

This patch changes the code which references
config_automatic_brightness_control to first check if the platform
has the system feature PackageManager.FEATURE_SENSOR_LIGHT. If it
doesn't, those resource values are set to false.

There is a similar patch to framework/base.

Change-Id: Ie2e850951dbab86b717f4703c64fec8d40abee7b
Signed-off-by: James Ketrenos <james.p.ketrenos@intel.com>
Reviewed-on: https://android.intel.com/316773
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ortiz, Samuel <samuel.ortiz@intel.com>
Tested-by: Ortiz, Samuel <samuel.ortiz@intel.com>
Reviewed-by: Mathew, Alwin <alwin.mathew@intel.com>
Reviewed-on: https://android.intel.com/322654
Reviewed-by: Bhat, Sudheendra S <sudheendra.s.bhat@intel.com>
Tested-by: Grape, Markus <markus.grape@intel.com>
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>